Proposed solution for issue #16, Below the fold images are sometimes being loaded.
After some digging I found that elements acting as the container MUST be visible when lazy image initializes. If the container has an animation on it and starts off hidden, it will not work and all images will load in container. This initial hidden state seems to screw up the calculations as the container element is not technically visible at the time lazy image is executed. This edge case causes the remaining to be a negative version of height.  Simply checking if remaining and height cancel each other out seemed to fix the edge case.